[ThirdParty] centralized prerequisite version config. Bump tinyusdz and cpptrace versions.
