+++ prerequisites.json
@@ -1,0 +1,5 @@
+{
+    "VulkanSDK": "1.4.313.0",
+    "ninja": "1.12.1",
+    "cmake": "3.30.5"
+}
